% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUSMap-package.R
\docType{package}
\name{BUSMap-package}
\alias{BUSMap-package}
\alias{BUSMap}
\title{Bayesian genotyping Uncertainty with Sequencing data for linkage Mapping (BUSMap)}
\description{
Package for constructing linkage maps using a Bayesian hierarchical hidden Markov model (HMM) 
in full-sib populations to obtain estimates of genetic map distances and associated uncertainty intervals.
}
\details{
\tabular{ll}{ Package: \tab BUSMap\cr Type: \tab Package\cr Version: \tab
0.2.0\cr Date: \tab 2020-04-29\cr License: \tab GPL 3\cr }
}
\author{
Timothy P. Bilton, Maintainer: Timothy P. Bilton
<timothy.bilton@agresearch.co.nz>
}
\keyword{package}
