% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMap.R
\name{computeMap}
\alias{computeMap}
\title{Construct linkage map using a Bayesian hierarchical hidden Markov model (HMM)}
\usage{
computeMap(ref, alt, OPGP, initval=NULL, iter=30000, burnin=5000, chains=3, seed=1, cores=chains)
}
\arguments{
\item{ref}{Non-negative integer matrix containing the read counts for the reference allele}

\item{alt}{Non-negative integer matrix containing the read counts for the alternate allele}

\item{OPGP}{Integer vector specifying the OPGPs}

\item{initval}{List giving the starting values for the MH algorithm value}

\item{iter}{Integer value of the total number of iterations in the MH algorithm for each chain}

\item{burnin}{Integer value of the number of interations in the burn-in/adaptive phase of the MH algorithm for each chain}

\item{chains}{Integer value giving the number of parallel chains to use}

\item{seed}{Numeric value giving the seed.}

\item{cores}{Integer value giving the number of cores for parallelization of the MH chains.
Must be non-negative and no more than \code{chains}.}
}
\description{
Function that implements the Metropolis-Hastings (MH) algorithm for obtaining posterior samples 
of the Bayesian hierarchical hidden Markov model (HMM) for linkage maps in full-sib families 
with high-throughput sequencing data.
}
\examples{

data(manuka)
ref <- manuka$ref
alt <- manuka$alt
OPGP <- manuka$OPGP

computeMap(ref, alt, OPGP)

}
\author{
Timothy P. Bilton
}
