% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMap.R
\name{computeMapSeq}
\alias{computeMapSeq}
\title{Construct a linkage map using high-throughput sequencing data}
\usage{
computeMapSeq(ref, alt, parHap, iter=30000, burnin=5000, chains=3, seed=1, cores=chains, initval=NULL)
}
\arguments{
\item{ref}{Non-negative integer matrix containing the read counts for the reference allele.}

\item{alt}{Non-negative integer matrix containing the read counts for the alternate allele.}

\item{parhap}{Matrix of parental haplotypes.}

\item{iter}{Integer value of the number of iterations in the MH algorithm (excluding burn-in period) for each chain.}

\item{burnin}{Integer value of the number of iterations in the burn-in/adaptive phase of the MH algorithm for each chain.}

\item{chains}{Integer value giving the number of parallel chains to use.}

\item{seed}{Numeric value giving the seed.}

\item{cores}{Integer value giving the number of cores for parallelization of the MH chains.}

\item{initval}{List giving the starting values for the MH algorithm value.}
}
\description{
Function that implements a Metropolis-Hastings (MH) algorithm for obtaining posterior samples 
of the Bayesian hierarchical hidden Markov model (HMM) for linkage maps in full-sib families 
with high-throughput sequencing data. Note that this function only constructs a linkage map for a single chromsome
in a single full-sib family.
}
\details{
The matrices for \code{ref} and \code{alt} must have the samples (or inidividuals) indexed by the rows
and the SNPs by the columns. Since this function is for full-sib families, only the read counts for the 
offspring should be included (the parental information is contained in \code{parhap}).

The \code{parhap} argument must be a matrix (4 rows and \eqn{M} columns, where \eqn{M} is the number of SNPs) containing the haplotypes for the parents. The entries 
of this matrix can be anything but there must be excatly two unique entires, one to denote the reference allele and
one to denote the alternate allele. The first two rows specify the haplotypes for the maternal parent and the third and
fourth rows specify the haplotypes for the paternal parent. See examples for one way to specify this matrix.

The \code{seed} argument allows for reproducibility and specifies the starting seed to when simulating random
values in the MCMC chains. 

The \code{initval} allows initial values for each chain to be specified. This must a list of vectors (each with length \eqn{2M+3}) with the number of elements
in the list being equal to the number of chains specified by \code{chains}. This argument should only be used by experienced users.
}
\examples{

#### Load Manuka data from Bilton et al., (2018)
data(manuka)
ref <- manuka$ref       # matrix of reference allele counts
alt <- manuka$alt       # matrix of alternate allele counts
parHap <- manuka$parHap # Matrix of parental haplotypes

computeMapSeq(ref, alt, parHap)

}
\author{
Timothy P. Bilton
}
